refactor(creative_agent): move output schemas to schemas.py
Extract the 12 Pydantic output-schema models (agent.py:57-546) into a new
creative_agent/schemas.py, mirroring the existing creative_eval/schemas.py
convention. agent.py re-imports them so they stay importable from
creative_agent.agent and keep object identity (output_schema kwargs and the
test_pipeline_structure identity asserts reference the same class objects).

Pure move: no schema field or agent behavior changed. agent.py 674 -> 476 lines.

Also folds Task 1b into the context-bloat file-splits plan doc.
