+++ creative_agent/agent.py
@@ -1,9 +1,7 @@
 import logging
 import warnings
-from typing import Literal
 
 from google.genai import types
-from pydantic import BaseModel, Field
 from google.adk.tools import google_search
 from google.adk.planners import BuiltInPlanner
 from google.adk.tools.agent_tool import AgentTool
@@ -16,6 +14,20 @@ from .config import config, INFRA_RETRY
 from . import callbacks
 from . import tools
 from . import prompts
+from .schemas import (  # noqa: F401
+    SearchQuery,
+    ResearchFeedback,
+    AdCopy,
+    AdCopyList,
+    FinalAdCopy,
+    FinalAdCopyList,
+    VisualConcept,
+    VisualConceptList,
+    VisualConceptCritique,
+    VisualConceptCritiqueList,
+    VisualConceptFinal,
+    VisualConceptFinalList,
+)
 from creative_eval.agent import creative_eval_agent
 
 
@@ -51,32 +63,6 @@ merge_parallel_insights = SequentialAgent(
     sub_agents=[parallel_planner_agent, merge_planners],
     description="Coordinates parallel research and synthesizes the results.",
 )
-
-
-# --- RESEARCH FEEDBACK SCHEMA --- #
-class SearchQuery(BaseModel):
-    """Model representing a specific search query for web search."""
-
-    search_query: str = Field(
-        description="A highly specific and targeted query for web search."
-    )
-
-
-class ResearchFeedback(BaseModel):
-    """Model for providing evaluation feedback on research quality."""
-
-    finding_type: Literal["Gap", "Opportunity"] = Field(
-        description="Evaluation result. 'Gap' if gathering missing data is most critical, 'Opportunity' if the remaining research should focus on exploring the nuances of the overlap/sentiment."
-    )
-
-    analysis_comment: str = Field(
-        description="Detailed explanation of the gap found OR the opportunity identified (max 3 sentences)."
-    )
-
-    follow_up_queries: list[SearchQuery] | None = Field(
-        default=None,
-        description="A list of specific, targeted follow-up search queries to either fill the identified gap or explore the highest-potential opportunity",
-    )
 
 
 combined_web_evaluator = Agent(
@@ -226,43 +212,6 @@ combined_research_pipeline = SequentialAgent(
 )
 
 
-# --- AD COPY SCHEMA ---
-class AdCopy(BaseModel):
-    """Model representing a single Ad Copy idea"""
-
-    id: int = Field(description="Numerical identifier; use values 1-10.")
-    tone_style: Literal[
-        "Humorous",
-        "Aspirational",
-        "Problem/Solution",
-        "Emotional/Authentic",
-        "Educational/Informative",
-        "Relatable/Meme-based",
-    ] = Field(description="Specify one of the required tones.")
-    headline: str = Field(description="A short, attention-grabbing Headline.")
-    body_text: str = Field(
-        description="2-3 sentences of concise and compelling ad copy."
-    )
-    trend_connection: str = Field(
-        description="A sentence explaining how this copy leverages or references the trend: {target_search_trends}."
-    )
-    audience_appeal_rationale: str = Field(
-        description="A brief, 1-sentence rationale for why this idea will appeal to the target audience, based on the research report."
-    )
-    social_caption: str = Field(
-        description="A candidate, short social media caption (e.g., for Instagram or TikTok video description)."
-    )
-
-
-class AdCopyList(BaseModel):
-    """Model for efficiently providing ad copy ideas for the critic agent to consume."""
-
-    ad_copies: list[AdCopy] | None = Field(
-        default=None,
-        description="A list of 10 initial ad copy ideas.",
-    )
-
-
 # --- AD COPY AGENT (DRAFT) ---
 ad_copy_drafter = Agent(
     model=build_gemini(config.worker_model),
@@ -288,47 +237,6 @@ ad_copy_drafter = Agent(
         callbacks.log_empty_turn_finish_reason,
     ],
 )
-
-
-# --- AD COPY SCHEMA FINAL ---
-class FinalAdCopy(BaseModel):
-    """Model representing a single Ad Copy idea"""
-
-    original_id: int = Field(description="Retain the original ID for traceability.")
-    tone_style: Literal[
-        "Humorous",
-        "Aspirational",
-        "Problem/Solution",
-        "Emotional/Authentic",
-        "Educational/Informative",
-        "Relatable/Meme-based",
-    ] = Field(description="The tone/style from the original idea (e.g., Humorous).")
-    headline: str = Field(description="The finalized, attention-grabbing Headline.")
-    body_text: str = Field(description="The finalized, concise and compelling ad copy.")
-    trend_connection: str = Field(
-        description="A sentence explaining how this copy leverages or references the trend: {target_search_trends}."
-    )
-    audience_appeal_rationale: str = Field(
-        description="A brief, 1-sentence rationale for target audience appeal."
-    )
-    social_caption: str = Field(
-        description="The finalized candidate social media caption."
-    )
-    call_to_action: str = Field(
-        description="A NEW, catchy, action-oriented phrase (e.g., 'Shop the drop now!')."
-    )
-    detailed_performance_rationale: str = Field(
-        description="A 2-3 sentence strategic critique explaining *why* this ad copy will perform well against the selection criteria."
-    )
-
-
-class FinalAdCopyList(BaseModel):
-    """Model for efficiently providing ad copy ideas for the critic agent to consume."""
-
-    ad_copies: list[FinalAdCopy] | None = Field(
-        default=None,
-        description="A list of the finalized ad copy ideas.",
-    )
 
 
 # --- AD COPY CRITIC AGENT ---
@@ -372,36 +280,6 @@ ad_creative_pipeline = SequentialAgent(
 )
 
 
-# --- VISUAL CONCEPT SCHEMA ---
-class VisualConcept(BaseModel):
-    """Model representing a single candidate visual concept."""
-
-    ad_copy_id: int = Field(
-        description="Retains the original ID for a direct link to the ad copy."
-    )
-    concept_name: str = Field(
-        description="A short, intuitive name for the visual concept."
-    )
-    trend_visual_link: str = Field(
-        description="A 1-sentence description of how the visual specifically incorporates the {target_search_trends}."
-    )
-    concept_summary: str = Field(
-        description="A 2-3 sentence explanation of the creative concept and its link to the ad copy's message."
-    )
-    image_generation_prompt: str = Field(
-        description="A draft prompt for image generation."
-    )
-
-
-class VisualConceptList(BaseModel):
-    """Model listing all initial visual concepts."""
-
-    visual_concepts: list[VisualConcept] | None = Field(
-        default=None,
-        description="A list of candidate visual concepts.",
-    )
-
-
 # --- VISUAL CONCEPT DRAFT AGENT ---
 visual_concept_drafter = Agent(
     model=build_gemini(config.worker_model),
@@ -424,35 +302,6 @@ visual_concept_drafter = Agent(
     output_key="visual_draft",
     after_model_callback=callbacks.log_empty_turn_finish_reason,
 )
-
-
-# --- VISUAL CONCEPT CRITIQUE SCHEMA ---
-class VisualConceptCritique(BaseModel):
-    """Model representing the critique of a single candidate visual concept."""
-
-    ad_copy_id: int = Field(
-        description="Retains the original ID for a direct link to the ad copy."
-    )
-    concept_name: str = Field(description="The original visual concept name.")
-    trend_visual_link: str = Field(description="The original trend link description.")
-    concept_summary: str = Field(
-        description="The original creative concept explanation."
-    )
-    image_generation_prompt: str = Field(
-        description="The FINAL, technically perfected, 100+ word, high-fidelity prompt."
-    )
-    critique_summary: str = Field(
-        description="A brief (1-2 sentence) summary of the key technical changes made to the prompt."
-    )
-
-
-class VisualConceptCritiqueList(BaseModel):
-    """Model listing all initial visual concepts."""
-
-    visual_concepts: list[VisualConceptCritique] | None = Field(
-        default=None,
-        description="A list of visual concept critiques.",
-    )
 
 
 # --- VISUAL CONCEPT CRITIQUE AGENT ---
@@ -479,53 +328,6 @@ visual_concept_critic = Agent(
     output_key="visual_concept_critique",
     after_model_callback=callbacks.log_empty_turn_finish_reason,
 )
-
-
-# --- VISUAL CONCEPT FINAL SCHEMA ---
-class VisualConceptFinal(BaseModel):
-    """Model representing a finalized visual concept."""
-
-    ad_copy_id: int = Field(
-        description="Retains the original ID for a direct link to the ad copy."
-    )
-    concept_name: str = Field(description="The finalized name of the visual concept.")
-    trend: str = Field(description="The trend referenced by this visual concept.")
-    trend_reference: str = Field(
-        description="How the visual concept relates to the target search trend"
-    )
-    markets_product: str = Field(
-        description="A brief explanation of how this markets the target product"
-    )
-    audience_appeal: str = Field(
-        description="A brief explanation for the target audience appeal."
-    )
-    selection_rationale: str = Field(
-        description="A brief rationale explaining why this visual concept was selected and why it will perform well"
-    )
-    headline: str = Field(
-        description="The final Headline text from the original ad copy."
-    )
-    social_caption: str = Field(
-        description="The final social media caption from the original ad copy."
-    )
-    call_to_action: str = Field(
-        description="The final Call-to-Action from the original ad copy."
-    )
-    concept_summary: str = Field(
-        description="A final, brief (2-3 sentence) summary of the combined ad copy and visual concept."
-    )
-    image_generation_prompt: str = Field(
-        description="The technically perfected, revised_image_generation_prompt."
-    )
-
-
-class VisualConceptFinalList(BaseModel):
-    """Model listing all finalized visual concepts."""
-
-    visual_concepts: list[VisualConceptFinal] | None = Field(
-        default=None,
-        description="A list of finalized visual concept.",
-    )
 
 
 # --- VISUAL CONCEPT FINAL AGENT ---
